pyproject.toml
[project]
name = "case-study-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.4",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "qwen-vl-utils>=0.0.14",
    "ruff>=0.15.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.0.0",
]
